+++ pom.xml
@@ -4,12 +4,12 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>ru.ezikvice.springotus</groupId>
-    <artifactId>homework8</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <artifactId>homework${homework.ver}</artifactId>
+    <version>${homework.ver}.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
-    <name>homework 8</name>
-    <description>Homework 8 for Otus Spring course</description>
+    <name>homework ${homework.ver}</name>
+    <description>Homework ${homework.ver} for Otus Spring course</description>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
@@ -19,6 +19,7 @@
     </parent>
 
     <properties>
+        <homework.ver>9</homework.ver>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <java.version>1.8</java.version>
