+++ src/main/java/ru/ezikvice/springotus/homework9/dao/BookRepository.java
@@ -1,7 +1,7 @@
-package ru.ezikvice.springotus.homework8.dao;
+package ru.ezikvice.springotus.homework9.dao;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import ru.ezikvice.springotus.homework8.domain.Book;
+import ru.ezikvice.springotus.homework9.domain.Book;
 
 public interface BookRepository extends MongoRepository<Book, Long> {
