+++ src/main/java/ru/ezikvice/springotus/homework9/service/GenreService.java
@@ -1,6 +1,6 @@
-package ru.ezikvice.springotus.homework8.service;
+package ru.ezikvice.springotus.homework9.service;
 
-import ru.ezikvice.springotus.homework8.domain.Genre;
+import ru.ezikvice.springotus.homework9.domain.Genre;
 
 public interface GenreService {
